moved rewrite code to its own package
